requirements.txt
# scripts/download_sra_data.py
requests>=2.28.0

# scripts/calculate_tetramer_frequencies.py (NumPy + Numba optional but recommended)
numpy>=1.24.0
numba>=0.59.0
lzma-mt>=0.1.4

# scripts/fit_tetramer_classifier.py
#scikit-learn>=1.3.0
# Pin version to avoid InconsistentVersionWarning: Trying to unpickle estimator ...
# https://scikit-learn.org/stable/model_persistence.html#security-maintainability-limitations
scikit-learn==1.8.0

# scripts/build_uc_cap_sequence_cache.py
pandas>=2.0.0
pyarrow>=14.0.0

# scripts/build_uc_cap_sequence_cache.py, scripts/run_uc_cap_pipeline.py, scripts/grid_uc_cap_pipeline.py
PyYAML>=6.0
